requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.1
pandas>=2.2.2
matplotlib>=3.9.0